Version 0.2 of the database

+++ database/iproject/iproject/Schema Objects/Schemas/dbo/Tables/Looptijd.table.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Looptijd]
 (
-	looptijd		INT					NOT NULL,
+	looptijd		TINYINT				NOT NULL,
 	actief			BIT					NOT NULL	DEFAULT CAST(1 AS BIT)
 
 	CONSTRAINT pk_looptijd				PRIMARY KEY (looptijd),
